requirements.txt
importlib-metadata>=6.0.0
langchain-google-vertexai==2.0.0
google-cloud-storage==2.18.2
google-cloud-aiplatform>=1.35.0
tiktoken
sentencepiece
pymupdf
google-cloud-storage
fastapi==0.115.0
uvicorn==0.30.6
numpy<2.0
python-multipart
langchain
langchain-community
langchain-huggingface
google-auth
pypdf==5.0.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
transformers
torch
protobuf
typing-extensions>=4.5.0
accelerate==0.30.0
huggingface_hub[hf_xet]

